Add middleware for access private page #JC19-355
