+++ apps/vanilla/src/pages/profile/profile.ts
@@ -1,3 +1,5 @@
+import { Middleware } from '../../namespaces/middleware';
 import { initHomeProfile } from '../../script/init';
 
+Middleware.isLoggedIn();
 initHomeProfile();
