+++ apps/vanilla/src/script/renderToUI.ts
@@ -5,7 +5,7 @@ import { getUserData } from '../services/getUserData';
  * Show login error message to UI.
  * @param errorList List of error message.
  */
-export function showErrorLogin(errorList: string[]): void {
+export function showErrorLogin(errorList: readonly string[]): void {
   const error = document.querySelector('.form__error-login');
   const errorMessage = errorList.join('<br>');
   if (error !== null) {
@@ -17,7 +17,7 @@ export function showErrorLogin(errorList: string[]): void {
  * Show register error message to UI.
  * @param errorList List of error message.
  */
-export function showErrorRegister(errorList: string[]): void {
+export function showErrorRegister(errorList: readonly string[]): void {
   const error = document.querySelector('.form__error-register');
   const errorMessage = errorList.join('<br>');
   if (error !== null) {
